pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "successor-quanta"
version = "0.1.0"
description = "Successor function experiments: quanta vs algorithm learning"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy",
    "wandb",
    "pyyaml",
    "matplotlib",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
